+++ R/get_rejection_rates.R
@@ -12,9 +12,13 @@
 #' the rejection rate of each test.
 #'
 #' @return
-#' The `rejection_rates` method for
-#' `power4test` objects returns
-#' a data frame with the number of
+#' The `rejection_rates` method returns
+#' a `rejection_rates_df` object,
+#' with a `print` method.
+#' If the input is a
+#' `power4test` object, it is
+#' a data-frame like object with the
+#' number of
 #' rows equal to the number of tests.
 #' Note that some tests, such as
 #' the test by [test_parameters()],
@@ -43,8 +47,11 @@
 #' a `power4test_by_n` object,
 #' or a `power4test_by_es` object.
 #'
-#' @param ... Optional arguments. Not
-#' used.
+#' @param ... Optional arguments. For
+#' the `print` method, these arguments
+#' will be passed to the `print` method
+#' of `data.frame` objects [print.data.frame()].
+#' Not used for other methods.
 #'
 #' @seealso [power4test()],
 #' [power4test_by_n()], and
@@ -144,6 +151,8 @@ rejection_rates.power4test <- function(object,
                     out1)
   }
   rownames(out2) <- NULL
+  class(out2) <- c("rejection_rates_df",
+                   class(out2))
   out2
 }
 
@@ -322,7 +331,11 @@ rbind_adv <- function(...) {
 #' @return
 #' The `rejection_rates` method for
 #' for `power4test_by_es` objects
-#' returns a data frame which is
+#' returns an object of the
+#' class `rejection_rates_df_by_es`,
+#' which is a subclass of
+#' `rejection_rates_df`.
+#' It is a data frame which is
 #' similar to the output of
 #' [rejection_rates()], with a
 #' column added for the effect size (`pop_es_name` and
@@ -342,14 +355,22 @@ rbind_adv <- function(...) {
 rejection_rates.power4test_by_es <- function(object,
                                              all_columns = FALSE,
                                              ...) {
-  rejection_rates_by_es(object_by_es = object,
-                        all_columns = all_columns)
+  out <- rejection_rates_by_es(object_by_es = object,
+                               all_columns = all_columns)
+  class(out) <- c("rejection_rates_df_by_es",
+                  "rejection_rates_df",
+                  class(out))
+  out
 }
 
 #' @return
 #' The `rejection_rates` method for
 #' for `power4test_by_n` objects
-#' returns a data frame which is
+#' returns an object of the
+#' class `rejection_rates_df_by_n`,
+#' which is a subclass of
+#' `rejection_rates_df`.
+#' It is a data frame which is
 #' similar to the output of
 #' for a `power4test` object, with a
 #' column `n` added for the sample size
@@ -367,6 +388,163 @@ rejection_rates.power4test_by_es <- function(object,
 rejection_rates.power4test_by_n <- function(object,
                                             all_columns = FALSE,
                                             ...) {
-  rejection_rates_by_n(object_by_n = object,
-                       all_columns = all_columns)
+  out <- rejection_rates_by_n(object_by_n = object,
+                              all_columns = all_columns)
+  class(out) <- c("rejection_rates_df_by_n",
+                  "rejection_rates_df",
+                  class(out))
+  out
+}
+
+#' @param x The `rejection_rates_df`
+#' object to be printed.
+#'
+#' @param digits The number of digits
+#' after the decimal place to be
+#' printed.
+#'
+#' @param annotation Logical. Whether
+#' additional notes will be printed.
+#'
+#' @param abbreviate_col_names Logical.
+#' Whether some column names will be
+#' abbreviated.
+#'
+#' @rdname rejection_rates
+#' @export
+print.rejection_rates_df <- function(x,
+                                     digits = 3,
+                                     annotation = TRUE,
+                                     abbreviate_col_names = TRUE,
+                                     ...) {
+  x0 <- x
+  class(x0) <- "data.frame"
+
+  # Handle special columns
+  if ("nvalid" %in% colnames(x0)) {
+    x0$nvalid <- as.character(x0$nvalid)
+  }
+  if ("nrep" %in% colnames(x0)) {
+    x0$nrep <- as.character(x0$nrep)
+  }
+
+  # In by_n
+  if ("n" %in% colnames(x0)) {
+    x0$n <- as.character(x0$n)
+  }
+
+  abbr_names <- c(pvalid = "p.v",
+                  nvalid = "n.v",
+                  reject_se = "r.se",
+                  reject_ci_lo = "r.cilo",
+                  reject_ci_hi = "r.cihi")
+
+  # Abbreviate column names
+  if (abbreviate_col_names) {
+    cnames <- colnames(x0)
+    for (i in seq_along(abbr_names)) {
+      cnames <- gsub(names(abbr_names)[i],
+                     abbr_names[i],
+                     cnames,
+                     fixed = TRUE)
+    }
+    colnames(x0) <- cnames
+  }
+
+  x1 <- format_num_cols(x0,
+                        digits = digits)
+  print(x1,
+        ...)
+
+  if (annotation) {
+    cat("Notes:\n")
+
+    if ("n" %in% colnames(x1)) {
+      catwrap(paste0("- ",
+                     "n",
+                     ": The sample size in a trial."),
+              exdent = 2)
+    }
+    if ("par" %in% colnames(x1)) {
+      catwrap(paste0("- ",
+                     "par",
+                     ": The parameter being varied."),
+              exdent = 2)
+    }
+    if ("es" %in% colnames(x1)) {
+      catwrap(paste0("- ",
+                     "es",
+                     ": The population value of 'par' in a trial."),
+              exdent = 2)
+    }
+    tmp <- ifelse(abbreviate_col_names,
+                  abbr_names["pvalid"],
+                  "pvalid")
+    if (tmp %in% colnames(x1)) {
+      catwrap(paste0("- ",
+                     tmp,
+                     ": The proportion of valid replications."),
+              exdent = 2)
+    }
+    tmp <- ifelse(abbreviate_col_names,
+                  abbr_names["nvalid"],
+                  "nvalid")
+    if (tmp %in% colnames(x1)) {
+      catwrap(paste0("- ",
+                     tmp,
+                     ": The number of valid replications."),
+              exdent = 2)
+    }
+    if ("nrep" %in% colnames(x1)) {
+      tmp <- "nrep"
+      catwrap(paste0("- ",
+                     tmp,
+                     ": The number of replications."),
+              exdent = 2)
+    }
+    if ("est" %in% colnames(x1)) {
+      catwrap(paste0("- est: The mean of the estimates in a test across replications."),
+              exdent = 2)
+    }
+    if ("cilo" %in% colnames(x1)) {
+      catwrap(paste0("- cilo, cihi: The mean of the lower/upper limits of ",
+                     "the confidence intervals in a test ",
+                     "across replications."),
+              exdent = 2)
+    }
+    if ("reject" %in% colnames(x1)) {
+      catwrap(paste0("- reject: The proportion of 'significant' replications, ",
+                     "that is, the rejection rate.",
+                     "If the null hypothesis is true, this is the Type I error rate. ",
+                     "If the null hypothesis is false, this is the power."),
+              exdent = 2)
+    }
+    tmp <- ifelse(abbreviate_col_names,
+                  abbr_names["reject_se"],
+                  "reject_se")
+    if (tmp %in% colnames(x1)) {
+      catwrap(paste0("- ",
+                     tmp,
+                     ": The standard error of the rejection rate, ",
+                     "based on normal approximation."),
+              exdent = 2)
+    }
+    tmp1 <- ifelse(abbreviate_col_names,
+                   abbr_names["reject_ci_lo"],
+                   "reject_ci_lo")
+    tmp2 <- ifelse(abbreviate_col_names,
+                   abbr_names["reject_ci_hi"],
+                   "reject_ci_hi")
+    if (tmp1 %in% colnames(x1)) {
+      tmp <- paste0(tmp1, ",", tmp2)
+      catwrap(paste0("- ",
+                     tmp,
+                     ": The confidence interval ",
+                     "of the rejection rate, based on normal approximation."),
+              exdent = 2)
+    }
+    catwrap(paste0("- Refer to the tests for the meanings of other columns."),
+            exdent = 2)
+  }
+
 }